requirements.txt
streamlit==1.16.0
PyPDF2==2.11.0
requests==2.31.0
python-dotenv==1.0.0
